+++ clients/spot/tests/unit/websocket-api/AccountApiTest.test.ts
@@ -103,6 +103,12 @@ describe('AccountApi', () => {
                         buyer: '0.00000030',
                         seller: '0.00000040',
                     },
+                    specialCommission: {
+                        maker: '0.01000000',
+                        taker: '0.02000000',
+                        buyer: '0.03000000',
+                        seller: '0.04000000',
+                    },
                     taxCommission: {
                         maker: '0.00000112',
                         taker: '0.00000114',