+++ clients/spot/tests/unit/websocket-api/GeneralApiTest.test.ts
@@ -164,12 +164,15 @@ describe('GeneralApi', () => {
                             minTrailingBelowDelta: 10,
                             maxTrailingBelowDelta: 2000,
                         },
+                        { filterType: 'MAX_NUM_ORDER_AMENDS', maxNumOrderAmends: 10 },
+                        { filterType: 'MAX_NUM_ORDER_LISTS', maxNumOrderLists: 20 },
                         { filterType: 'EXCHANGE_MAX_NUM_ORDERS', maxNumOrders: 1000 },
                         { filterType: 'EXCHANGE_MAX_NUM_ALGO_ORDERS', maxNumAlgoOrders: 200 },
                         {
                             filterType: 'EXCHANGE_MAX_NUM_ICEBERG_ORDERS',
                             maxNumIcebergOrders: 10000,
                         },
+                        { filterType: 'EXCHANGE_MAX_NUM_ORDER_LISTS', maxNumOrderLists: 20 },
                     ],
                     symbols: [
                         {
@@ -192,6 +195,7 @@ describe('GeneralApi', () => {
                             allowTrailingStop: true,
                             cancelReplaceAllowed: true,
                             amendAllowed: false,
+                            pegInstructionsAllowed: true,
                             isSpotTradingAllowed: true,
                             isMarginTradingAllowed: true,
                             filters: [
@@ -253,6 +257,8 @@ describe('GeneralApi', () => {
                                     minTrailingBelowDelta: 10,
                                     maxTrailingBelowDelta: 2000,
                                 },
+                                { filterType: 'MAX_NUM_ORDER_AMENDS', maxNumOrderAmends: 10 },
+                                { filterType: 'MAX_NUM_ORDER_LISTS', maxNumOrderLists: 20 },
                                 { filterType: 'EXCHANGE_MAX_NUM_ORDERS', maxNumOrders: 1000 },
                                 {
                                     filterType: 'EXCHANGE_MAX_NUM_ALGO_ORDERS',
@@ -261,6 +267,10 @@ describe('GeneralApi', () => {
                                 {
                                     filterType: 'EXCHANGE_MAX_NUM_ICEBERG_ORDERS',
                                     maxNumIcebergOrders: 10000,
+                                },
+                                {
+                                    filterType: 'EXCHANGE_MAX_NUM_ORDER_LISTS',
+                                    maxNumOrderLists: 20,
                                 },
                             ],
                             permissions: [],